fix(issues): move lifecycle buttons back to the drawer top toolbar [LAT-634]
Ignore/Resolve had drifted into the title/description row after the
session-drawer refactor. Extract them into a reusable
`IssueLifecycleActions` so the standalone issue drawer renders them in
`rightActions` (next to next/prev) and the session drawer renders them
in `rightActions` (next to "View session") when an issue is open.
